Create 4.6.3
Declare a Boolean variable named matchesCond. Then, read integer numInput from input representing the number of integers to be read next. Use a loop to read the remaining integers from input. If all numInput integers are equal to 1, assign matchesCond with true. Otherwise, assign matchesCond with false.

Code at the end of main() outputs "All match 1" if matchesCond is true or "Has non-matching value(s)" if matchesCond is false.

Ex: If the input is:

3
1 1 1
then the output is:

All match 1