4.7.1
A website requires that passwords only contain letters or numbers. For each character in password that is not a letter or number, replace the character with 'x'.

Ex: If the input is 2k mr, then the output is:

Valid password: 2kxmr

Note:

isalpha() returns true if a character is a letter, and false otherwise. Ex: isalpha('a') returns true. isalpha('8') returns false.
isdigit() returns true if a character is a digit, and false otherwise. Ex: isdigit('8') returns true. isdigit('a') returns false.