Create 4.7.1: Password requirements.
Declare a Boolean variable named passwordValid. Use passwordValid to output "Valid" if keyString contains no more than 6 digits and keyString's length is less than or equal to 7. Otherwise, output "Invalid".

Ex: If the input is 798w203, then the output is:

Valid

Ex: If the input is VRtFMSDG, then the output is:

Invalid

Note: isdigit() returns true if a character is a digit, and false otherwise. Ex: isdigit('8') returns true. isdigit('a') returns false.